code refactored...
