+++ src/services/general/index.js
@@ -1,0 +1,3 @@
+export const general = async (request, response) => {
+  return { success: true };
+};
